requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
html2text
mysql-connector-python
pytest
python-dotenv
pytest-mock
tabulate